+++ tests/integration/mtp/test_agent_permission_chain.py
@@ -1,0 +1,325 @@
+"""Agent 权限链路集成测试（Profile → Prompt → Koakuma 运行时拦截）。"""
+
+from types import MethodType
+from unittest.mock import Mock, patch
+from uuid import uuid4
+
+import pytest
+
+from hivememory.agent_runtime.models import MTPExecutionContext
+from hivememory.agent_runtime.mtp.runtime import KoakumaRuntime
+from hivememory.core.models import (
+    OMNI_DOLL_PROFILE,
+    AgentProfile,
+    IndexLayer,
+    MemoryAtom,
+    MemoryType,
+    MetaData,
+    PayloadLayer,
+)
+from hivememory.core.mtp.exceptions import (
+    AliasNotFoundError,
+    InvalidArgumentError,
+    PermissionDeniedError,
+)
+from hivememory.patchouli.runtime.core import PatchouliRuntime
+from hivememory.prompts.mtp import MTPPromptBuilder
+
+
+def _make_profile_atom(
+    agent_id: str,
+    allowed_verbs: list,
+    allowed_tools: list,
+) -> MemoryAtom:
+    """构建 AGENT_PROFILE 类型的 MemoryAtom"""
+    return MemoryAtom(
+        id=uuid4(),
+        meta=MetaData(
+            user_id="system",
+            source_agent_id="system",
+            visibility="PUBLIC",
+        ),
+        index=IndexLayer(
+            alias=agent_id,
+            title=f"Agent {agent_id}",
+            summary=f"Profile for {agent_id}",
+            tags=["agent", "profile"],
+            memory_type=MemoryType.AGENT_PROFILE,
+        ),
+        payload=PayloadLayer(
+            content=f"You are {agent_id}.",
+            artifacts={
+                "agent_config": {
+                    "model_name": "gpt-4",
+                    "temperature": 0.7,
+                    "allowed_mtp_verbs": allowed_verbs,
+                    "allowed_sys_tools": allowed_tools,
+                }
+            },
+        ),
+    )
+
+
+def _create_runtime_with_koakuma():
+    """创建带 Koakuma 的 PatchouliRuntime"""
+    with patch.object(PatchouliRuntime, "_init_infrastructure"), \
+         patch.object(PatchouliRuntime, "_build_engines", return_value={}), \
+         patch.object(PatchouliRuntime, "_register_services"):
+
+        mock_patchouli_config = Mock()
+        mock_shared_config = Mock()
+
+        runtime = PatchouliRuntime(patchouli_config=mock_patchouli_config, shared_config=mock_shared_config)
+        runtime.storage = Mock()
+
+        def _get_agent_profile(agent_alias: str):
+            if not agent_alias or agent_alias in ("default", "omni_doll"):
+                return OMNI_DOLL_PROFILE
+            atom = runtime.storage.get_memory_by_alias(agent_alias)
+            if atom is None:
+                raise AliasNotFoundError(
+                    message_key="mtp.call.profile_not_found",
+                    params={"agent_alias": agent_alias},
+                )
+            profile = AgentProfile.from_atom(atom)
+            if profile is None:
+                raise InvalidArgumentError(
+                    message_key="mtp.call.profile_invalid",
+                    params={"agent_alias": agent_alias},
+                )
+            return profile
+
+        def _get_mtp_prompt(self, profile=None):
+            builder = MTPPromptBuilder(
+                language="en",
+                include_demo=True,
+                include_error_handling=True,
+                allowed_verbs=getattr(profile, "allowed_mtp_verbs", None),
+                allowed_runtime_tools=getattr(profile, "allowed_sys_tools", None),
+            )
+            return builder.build()
+
+        runtime.storage.get_agent_profile = Mock(side_effect=_get_agent_profile)
+        runtime.get_mtp_prompt = MethodType(_get_mtp_prompt, runtime)
+
+        # 创建真实的 Koakuma 实例
+        koakuma = KoakumaRuntime(bus=None, config=None, alias_resolver=Mock())
+        runtime._services = {"koakuma": koakuma}
+
+        return runtime, koakuma
+
+
+@pytest.mark.asyncio
+class TestPromptToKoakumaEnforcement:
+    """Prompt → Koakuma 运行时拦截链路测试"""
+
+    async def test_koakuma_enforces_verb_permission(self):
+        """Koakuma 运行时拦截越权动词"""
+        kernel, koakuma = _create_runtime_with_koakuma()
+
+        # 设置限制性 profile
+        profile_atom = _make_profile_atom(
+            agent_id="reader_only",
+            allowed_verbs=["READ", "SEARCH"],
+            allowed_tools=["sys_clock"],
+        )
+        kernel.storage.get_memory_by_alias = Mock(return_value=profile_atom)
+
+        profile = kernel.storage.get_agent_profile("reader_only")
+        context = MTPExecutionContext(agent_profile=profile)
+
+        # 允许的动词应该通过
+        koakuma._check_verb_permission("READ", context=context)
+        koakuma._check_verb_permission("SEARCH", context=context)
+
+        # 禁止的动词应该被拦截
+        with pytest.raises(PermissionDeniedError):
+            koakuma._check_verb_permission("WRITE", context=context)
+
+        with pytest.raises(PermissionDeniedError):
+            koakuma._check_verb_permission("UPDATE", context=context)
+
+    async def test_koakuma_enforces_tool_permission(self):
+        """Koakuma 运行时拦截越权工具"""
+        kernel, koakuma = _create_runtime_with_koakuma()
+
+        profile_atom = _make_profile_atom(
+            agent_id="safe_agent",
+            allowed_verbs=["READ", "RUN"],
+            allowed_tools=["sys_clock"],
+        )
+        kernel.storage.get_memory_by_alias = Mock(return_value=profile_atom)
+
+        profile = kernel.storage.get_agent_profile("safe_agent")
+        context = MTPExecutionContext(agent_profile=profile)
+
+        # 允许的工具应该通过
+        koakuma._check_tool_permission("sys_clock")
+
+        # 禁止的工具应该被拦截
+        with pytest.raises(PermissionDeniedError):
+            koakuma._check_tool_permission("sys_write_file", context=context)
+
+        with pytest.raises(PermissionDeniedError):
+            koakuma._check_tool_permission("sys_bash_exec", context=context)
+
+
+@pytest.mark.asyncio
+class TestEndToEndPermissionChain:
+    """端到端权限链路测试"""
+
+    async def test_reviewer_profile_blocks_write(self):
+        """Reviewer profile: prompt 不显示 WRITE，运行时拦截 WRITE"""
+        kernel, koakuma = _create_runtime_with_koakuma()
+
+        # 加载 reviewer profile
+        reviewer_atom = _make_profile_atom(
+            agent_id="reviewer_doll",
+            allowed_verbs=["READ", "SEARCH"],
+            allowed_tools=["sys_clock"],
+        )
+        kernel.storage.get_memory_by_alias = Mock(return_value=reviewer_atom)
+
+        profile = kernel.storage.get_agent_profile("reviewer_doll")
+
+        # 1. Prompt 层：WRITE 不应该出现
+        mtp_prompt = kernel.get_mtp_prompt(profile=profile)
+        assert "- WRITE:" not in mtp_prompt
+        assert "- UPDATE:" not in mtp_prompt
+
+        # 2. Runtime 层：Koakuma 拦截 WRITE
+        context = MTPExecutionContext(agent_profile=profile)
+
+        with pytest.raises(PermissionDeniedError) as exc_info:
+            koakuma._check_verb_permission("WRITE", context=context)
+
+        assert "WRITE" in str(exc_info.value)
+
+    async def test_coder_profile_allows_write(self):
+        """Coder profile: prompt 显示 WRITE，运行时允许 WRITE"""
+        kernel, koakuma = _create_runtime_with_koakuma()
+
+        # 加载 coder profile
+        coder_atom = _make_profile_atom(
+            agent_id="coder_doll",
+            allowed_verbs=["READ", "SEARCH", "WRITE", "RUN"],
+            allowed_tools=["sys_clock", "sys_read_file", "sys_write_file"],
+        )
+        kernel.storage.get_memory_by_alias = Mock(return_value=coder_atom)
+
+        profile = kernel.storage.get_agent_profile("coder_doll")
+
+        # 1. Prompt 层：WRITE 应该出现
+        mtp_prompt = kernel.get_mtp_prompt(profile=profile)
+        assert "- WRITE:" in mtp_prompt
+        assert "- RUN:" in mtp_prompt
+        assert "sys_write_file" in mtp_prompt
+
+        # 2. Runtime 层：Koakuma 允许 WRITE
+        context = MTPExecutionContext(agent_profile=profile)
+
+        # 应该不抛异常
+        koakuma._check_verb_permission("WRITE", context=context)
+        koakuma._check_verb_permission("RUN", context=context)
+        koakuma._check_tool_permission("sys_write_file", context=context)
+
+    async def test_no_profile_allows_all_operations(self):
+        """无 profile 时：prompt 显示全部，运行时允许全部"""
+        kernel, koakuma = _create_runtime_with_koakuma()
+
+        # 1. Prompt 层：全部动词和工具都应该出现
+        mtp_prompt = kernel.get_mtp_prompt(profile=None)
+        assert "- SEARCH:" in mtp_prompt
+        assert "- READ:" in mtp_prompt
+        assert "- WRITE:" in mtp_prompt
+        assert "- UPDATE:" in mtp_prompt
+        assert "- RUN:" in mtp_prompt
+
+        # 2. Runtime 层：Koakuma 允许全部（无 profile）
+        context = MTPExecutionContext()
+
+        # 应该不抛异常
+        koakuma._check_verb_permission("WRITE", context=context)
+        koakuma._check_verb_permission("UPDATE", context=context)
+        koakuma._check_verb_permission("RUN", context=context)
+        koakuma._check_tool_permission("sys_write_file", context=context)
+        koakuma._check_tool_permission("sys_bash_exec", context=context)
+
+
+@pytest.mark.asyncio
+class TestSecurityScenarios:
+    """安全场景测试"""
+
+    async def test_prompt_injection_cannot_bypass_runtime(self):
+        """Prompt 注入无法绕过运行时拦截"""
+        kernel, koakuma = _create_runtime_with_koakuma()
+
+        # 设置限制性 profile
+        profile_atom = _make_profile_atom(
+            agent_id="restricted",
+            allowed_verbs=["READ"],
+            allowed_tools=[],
+        )
+        kernel.storage.get_memory_by_alias = Mock(return_value=profile_atom)
+
+        profile = kernel.storage.get_agent_profile("restricted")
+        context = MTPExecutionContext(agent_profile=profile)
+
+        # 即使 LLM 幻觉输出了 WRITE 指令，运行时也会拦截
+        with pytest.raises(PermissionDeniedError):
+            koakuma._check_verb_permission("WRITE", context=context)
+
+    async def test_tool_permission_exact_match(self):
+        """工具权限精确匹配（防止前缀攻击）"""
+        kernel, koakuma = _create_runtime_with_koakuma()
+
+        profile_atom = _make_profile_atom(
+            agent_id="limited",
+            allowed_verbs=["RUN"],
+            allowed_tools=["sys_clock"],
+        )
+        kernel.storage.get_memory_by_alias = Mock(return_value=profile_atom)
+
+        profile = kernel.storage.get_agent_profile("limited")
+        context = MTPExecutionContext(agent_profile=profile)
+
+        # sys_clock 应该通过
+        koakuma._check_tool_permission("sys_clock", context=context)
+
+        # sys_clock_evil 不应该通过（不是前缀匹配）
+        with pytest.raises(PermissionDeniedError):
+            koakuma._check_tool_permission("sys_clock_evil", context=context)
+
+    async def test_profile_switch_updates_permissions(self):
+        """Profile 切换正确更新权限"""
+        kernel, koakuma = _create_runtime_with_koakuma()
+
+        # 第一个 profile：限制性
+        restrictive_atom = _make_profile_atom(
+            agent_id="restrictive",
+            allowed_verbs=["READ"],
+            allowed_tools=[],
+        )
+        kernel.storage.get_memory_by_alias = Mock(return_value=restrictive_atom)
+
+        profile1 = kernel.storage.get_agent_profile("restrictive")
+        context1 = MTPExecutionContext(agent_profile=profile1)
+
+        # WRITE 应该被拒绝
+        with pytest.raises(PermissionDeniedError):
+            koakuma._check_verb_permission("WRITE", context=context1)
+
+        # 切换到第二个 profile：宽松
+        permissive_atom = _make_profile_atom(
+            agent_id="permissive",
+            allowed_verbs=["READ", "WRITE"],
+            allowed_tools=["sys_write_file"],
+        )
+        kernel.storage.get_memory_by_alias = Mock(return_value=permissive_atom)
+
+        profile2 = kernel.storage.get_agent_profile("permissive")
+        context2 = MTPExecutionContext(agent_profile=profile2)
+
+        # WRITE 现在应该通过
+        koakuma._check_verb_permission("WRITE", context=context2)
+        koakuma._check_tool_permission("sys_write_file", context=context2)
