+++ tests/integration/mtp/test_koakuma_write_e2e.py
@@ -1,0 +1,128 @@
+"""Koakuma WRITE 指令链路集成测试。"""
+
+import pytest
+from unittest.mock import MagicMock
+
+from hivememory.core.models import (
+    Identity,
+    IndexLayer,
+    MemoryAtom,
+    MemoryType,
+    MetaData,
+    PayloadLayer,
+)
+from hivememory.agent_runtime.models import MTPExecutionContext
+from hivememory.agent_runtime.mtp.runtime import KoakumaRuntime
+from hivememory.system.config import KoakumaConfig
+
+from .conftest import (
+    make_koakuma_runtime,
+    make_mock_bus,
+    normalize_worker_agent_mtp_output,
+)
+
+
+@pytest.fixture
+def identity() -> Identity:
+    return Identity(user_id="test_user", agent_id="test_agent", session_id="test_session")
+
+
+@pytest.fixture
+def sample_memory(identity) -> MemoryAtom:
+    return MemoryAtom(
+        meta=MetaData(user_id=identity.user_id, source_agent_id=identity.agent_id, session_id=None, confidence_score=1.0),
+        index=IndexLayer(title="Fix CORS", summary="修复 CORS 跨域问题，端口从 8080 改为 9090", tags=["cors"], memory_type=MemoryType.FACT),
+        payload=PayloadLayer(content="端口从 8080 改为 9090"),
+    )
+
+
+async def _intercept_and_execute(koakuma: KoakumaRuntime, assistant_text: str, context=None):
+    return await koakuma.intercept_and_execute(
+        normalize_worker_agent_mtp_output(assistant_text),
+        context=context,
+    )
+
+
+class TestKoakumaWriteE2E:
+    """通过 MTP 指令验证 Koakuma WRITE 完整链路"""
+
+    @pytest.fixture
+    def write_koakuma(self, sample_memory) -> KoakumaRuntime:
+        mock_librarian = MagicMock()
+        mock_librarian.handle_write_signal.return_value = [sample_memory]
+
+        bus = make_mock_bus()
+        koakuma = make_koakuma_runtime(bus, KoakumaConfig())
+        koakuma.context = MTPExecutionContext(identity=Identity(user_id="test_user"))
+        return koakuma
+
+    @pytest.mark.asyncio
+    async def test_write_basic(self, write_koakuma):
+        agent_text = '⟪ WRITE | * | content="端口从 8080 改为 9090" reason="修复 CORS"'
+        result = await _intercept_and_execute(write_koakuma, agent_text, context=write_koakuma.context)
+
+        assert result is not None
+        assert result.success
+
+        pending = write_koakuma.pending_runtime.get(result.pending_alias)
+        assert pending is not None
+        focus = pending.focus
+        assert focus.content == "端口从 8080 改为 9090"
+        assert focus.reason == "修复 CORS"
+        assert pending.identity.user_id == "test_user"
+
+    @pytest.mark.asyncio
+    async def test_write_with_title(self, write_koakuma):
+        agent_text = '⟪ WRITE | * | title="Fix CORS" content="端口改为 9090" reason="修复"'
+        result = await _intercept_and_execute(write_koakuma, agent_text, context=write_koakuma.context)
+
+        assert result is not None
+        pending = write_koakuma.pending_runtime.get(result.pending_alias)
+        assert pending is not None
+        focus = pending.focus
+        assert focus.title == "Fix CORS"
+
+    @pytest.mark.asyncio
+    async def test_write_missing_content(self, write_koakuma):
+        agent_text = '⟪ WRITE | * | reason="no content"'
+        result = await _intercept_and_execute(write_koakuma, agent_text, context=write_koakuma.context)
+
+        assert result is not None
+        assert result.pending_alias is None
+
+    @pytest.mark.asyncio
+    async def test_write_response_contains_ack(self, write_koakuma):
+        agent_text = '⟪ WRITE | * | content="test content"'
+        result = await _intercept_and_execute(write_koakuma, agent_text, context=write_koakuma.context)
+
+        assert result is not None
+        assert result.pending_alias is not None
+        assert "pending atom" in result.response_content
+        assert result.pending_alias in result.response_content
+        assert "ack" in result.formatted_response.lower()
+
+    @pytest.mark.asyncio
+    async def test_write_deferred_capture_always_ack(self):
+        """v3.0 延迟捕获: WRITE 在 Koakuma 层始终返回 ACK，实际执行延迟到 payload 提交"""
+        bus = make_mock_bus()
+        koakuma = make_koakuma_runtime(bus, KoakumaConfig())
+        from hivememory.core.models import RuntimeScope
+
+        context = MTPExecutionContext(
+            identity=Identity(user_id="test_user"),
+            runtime_scope=RuntimeScope(
+                run_id="run_write_test",
+                frame_id="frame_main_write",
+            ),
+        )
+
+        agent_text = '⟪ WRITE | * | content="test"'
+        result = await _intercept_and_execute(koakuma, agent_text, context=context)
+
+        assert result is not None
+        assert result.success
+        pending = koakuma.pending_runtime.get(result.pending_alias)
+        assert pending is not None
+        assert pending.focus.content == "test"
+        assert pending.runtime_scope.run_id == "run_write_test"
+        assert pending.runtime_scope.frame_id == "frame_main_write"
