+++ tests/unit/agent_runtime/mtp/test_write_chain.py
@@ -21,11 +21,6 @@ from datetime import datetime
 
 from hivememory.core.models import (
     Identity,
-    IndexLayer,
-    MemoryAtom,
-    MemoryType,
-    MetaData,
-    PayloadLayer,
     StreamMessage,
     StreamMessageType,
     WriteFocus,
@@ -39,10 +34,6 @@ from hivememory.engines.generation.models import (
 )
 from hivememory.engines.perception.models import FlushReason
 from hivememory.engines.generation.engine import MemoryGenerationEngine
-from hivememory.agent_runtime.mtp.runtime import KoakumaRuntime
-from hivememory.agent_runtime.models import MTPExecutionContext
-from hivememory.system.config import KoakumaConfig
-from hivememory.core.mtp import MTPResponseStatus
 
 
 # ========== Fixtures ==========
@@ -72,47 +63,11 @@ def sample_context(sample_messages, identity) -> GenerationContext:
         ]
     )
 
-@pytest.fixture
-def sample_memory(identity) -> MemoryAtom:
-    return MemoryAtom(
-        meta=MetaData(user_id=identity.user_id, source_agent_id=identity.agent_id, session_id=None, confidence_score=1.0),
-        index=IndexLayer(title="Fix CORS", summary="修复 CORS 跨域问题，端口从 8080 改为 9090", tags=["cors"], memory_type=MemoryType.FACT),
-        payload=PayloadLayer(content="端口从 8080 改为 9090"),
-    )
-
-
-@pytest.fixture
-def sample_draft() -> ExtractedMemoryDraft:
-    return ExtractedMemoryDraft(
-        title="Fix CORS",
-        summary="端口改为 9090",
-        tags=["cors", "nginx"],
-        memory_type="FACT",
-        content="端口从 8080 改为 9090",
-        confidence_score=1.0,
-        has_value=True,
-        alias_suffix="fix_cors",
-    )
-
-
 def _mock_mid_term():
     mid_term = MagicMock()
     mid_term.search = AsyncMock(return_value=[])
     mid_term.upsert = AsyncMock()
     return mid_term
-
-
-async def _execute_mtp(koakuma: KoakumaRuntime, text: str, context=None):
-    return await koakuma.execute_mtp(text, context=context)
-
-
-async def _intercept_and_execute(koakuma: KoakumaRuntime, assistant_text: str, context=None):
-    from .conftest import normalize_worker_agent_mtp_output
-
-    return await koakuma.intercept_and_execute(
-        normalize_worker_agent_mtp_output(assistant_text),
-        context=context,
-    )
 
 
 # ========== Test 1: WriteFocus Model ==========
@@ -304,95 +259,6 @@ class TestActiveWriteBoundary:
         assert request.is_write
         assert request.write_focus is focus
         assert request.update_focus is None
-
-
-# ========== Test 7: Koakuma WRITE E2E ==========
-
-class TestKoakumaWriteE2E:
-    """通过 MTP 指令验证 Koakuma WRITE 完整链路"""
-
-    @pytest.fixture
-    def write_koakuma(self, sample_memory) -> KoakumaRuntime:
-        mock_librarian = MagicMock()
-        mock_librarian.handle_write_signal.return_value = [sample_memory]
-
-        from .conftest import make_koakuma_runtime, make_mock_bus
-        bus = make_mock_bus()
-        koakuma = make_koakuma_runtime(bus, KoakumaConfig())
-        koakuma.context = MTPExecutionContext(identity=Identity(user_id="test_user"))
-        return koakuma
-
-    @pytest.mark.asyncio
-    async def test_write_basic(self, write_koakuma):
-        agent_text = '⟪ WRITE | * | content="端口从 8080 改为 9090" reason="修复 CORS"'
-        result = await _intercept_and_execute(write_koakuma, agent_text, context=write_koakuma.context)
-
-        assert result is not None
-        assert result.success
-
-        pending = write_koakuma.pending_runtime.get(result.pending_alias)
-        assert pending is not None
-        focus = pending.focus
-        assert focus.content == "端口从 8080 改为 9090"
-        assert focus.reason == "修复 CORS"
-        assert pending.identity.user_id == "test_user"
-
-    @pytest.mark.asyncio
-    async def test_write_with_title(self, write_koakuma):
-        agent_text = '⟪ WRITE | * | title="Fix CORS" content="端口改为 9090" reason="修复"'
-        result = await _intercept_and_execute(write_koakuma, agent_text, context=write_koakuma.context)
-
-        assert result is not None
-        pending = write_koakuma.pending_runtime.get(result.pending_alias)
-        assert pending is not None
-        focus = pending.focus
-        assert focus.title == "Fix CORS"
-
-    @pytest.mark.asyncio
-    async def test_write_missing_content(self, write_koakuma):
-        agent_text = '⟪ WRITE | * | reason="no content"'
-        result = await _intercept_and_execute(write_koakuma, agent_text, context=write_koakuma.context)
-
-        assert result is not None
-        assert result.pending_alias is None
-
-    @pytest.mark.asyncio
-    async def test_write_response_contains_ack(self, write_koakuma):
-        agent_text = '⟪ WRITE | * | content="test content"'
-        result = await _intercept_and_execute(write_koakuma, agent_text, context=write_koakuma.context)
-
-        assert result is not None
-        assert result.pending_alias is not None
-        assert "pending atom" in result.response_content
-        assert result.pending_alias in result.response_content
-        assert "ack" in result.formatted_response.lower()
-
-    @pytest.mark.asyncio
-    async def test_write_deferred_capture_always_ack(self):
-        """v3.0 延迟捕获: WRITE 在 Koakuma 层始终返回 ACK，实际执行延迟到 payload 提交"""
-        from .conftest import make_koakuma_runtime, make_mock_bus
-        bus = make_mock_bus()
-        koakuma = make_koakuma_runtime(bus, KoakumaConfig())
-        from hivememory.core.models import RuntimeScope
-
-        context = MTPExecutionContext(
-            identity=Identity(user_id="test_user"),
-            runtime_scope=RuntimeScope(
-                run_id="run_write_test",
-                frame_id="frame_main_write",
-            ),
-        )
-
-        agent_text = '⟪ WRITE | * | content="test"'
-        result = await _intercept_and_execute(koakuma, agent_text, context=context)
-
-        assert result is not None
-        assert result.success
-        pending = koakuma.pending_runtime.get(result.pending_alias)
-        assert pending is not None
-        assert pending.focus.content == "test"
-        assert pending.runtime_scope.run_id == "run_write_test"
-        assert pending.runtime_scope.frame_id == "frame_main_write"
 
 
 # ========== Test 8: Active Flush Reason Removed ==========
